add encoded-word support
Add encoded-word support.  The 'decodeEncodedWords' function is
provided for locating and decoding =?...?= sequences within strings.
The Q and B encodings are supported.  This is not a strict
interpretation of RFC 2047 - whereas encoded-word is defined in the
grammar, 'decodeEncodedWords' does post-parsing search and replace.

Supported charsets are UTF-8, ISO-8859-1 and ASCII.  Windows-1252
and Big5 were found within my corpus of email so should probably be
prioritised.