implement message charset decoding
After content transfer decoding, the body is a ByteString whose text
encoding is indicated in the 'charset' parameter of the Content-Type
header.  Text parts should be decoded according to that indication.

To implement this safely, add a phantom type parameter to 'Message'
for a "context".  This is used to track what state the message is in
- that is, is it a wire-formatted ByteString, a transfer-decoded
ByteString, or something else?  Add type synonyms for common
combinations of encoding state and payload type.

Implement optics for performing the conversions between these states
(only decoding, so far).  Initially supported charsets are UTF-8,
US-ASCII and ISO-8859-1.

Also rename messageHeaders -> headers and messageBody -> body.