fix parameter decoding when no charset specified
Decoding fails when the Content-Type header did not specify a
'charset' parameter, or when the Content-Disposition 'filename'
parameter did not specify a charset via the extension mechanism
defined in RFC 2231.

In both of these cases, the default character set is declared or
implied to be US-ASCII.  Therefore, all use cases of the HasCharset
class either require a charset to be specified, or have a default
charset.  The partiality of the 'charsetName' optic in the
'HasCharset' type class goes away, as does the 'CharsetUnspecified'
case of the 'CharsetError' data type.