update parameter optics to support writing
Change 'Parameters' from a type synonym to a newtype and add
instances of Ixed and At.  The optics now support writing.  Writing
a ParameterValue erases all "related keys" from the underlying and
writes the value as 1 or more raw parameters (>1 if continuation
required due to length).

Note that continuations are not yet implemented on the write site;
long parameters are incorrectly written as-is.