encode parameters using quoted-string where necessary
Some ASCII characters cannot appear in the plain parameter encoding.
For compatibility reasons it is desirable to avoid the extended
encoding where possible, so when plain encoding is not capable, use
the quoted-string encoding where possible.

A semilattice structure ('Max ParameterEncoding') is used to
determine the best encoding.  This "default" is overriden (forced to
Extended) when the charset or language is specified.

Add a number of new tests, including a roundtrip property.