Provide an optic to retrieve a value setting the 'References' field
RFC5322 - 3.6.4 specifies identification fields used when the user replies to a
message. This patch adds a getter which might return a ByteString in the effect
the rules defined in 3.6.4 return the necessary header values from the original
message. The ByteString can be used to set the 'references' field value in the
replied message header.