avoid having to set MIME-Version header
Make 'renderMessage' and 'buildMessage' automatically add the
"MIME-Version: 1.0" header.  Update places where we were doing that,
to not do it (because it is now done automatically).

Update Eq instances to ignore the MIME-Version header (to satisfy
roundtrip properties).  Because we only want to do this for 'MIME'
body type, and not any old message, define the 'EqMessage' class for
specifying how 'Message' at a particular body type can be compared.
Redefine (==) for 'Message s' as

  instance EqMessage a => Eq (Message s a)

and define the instance for 'MIME', which ignores the "MIME-Version"
header.