MIME: add FailedParse constructor
When parsing a MIME message, if a subpart fails to parse, instead of
failing the whole parse, add a dedicated constructor for the bit
that failed.  Also add the 'MIMEParseError' data type for specifying
what kind of failure occurred (boundary not specified, failed to
parse multipart structure, failed to parse part as Message).