support parsing messages with no body
RFC 5322 and its predecessors define message as:

   message         =   (fields / obs-fields)
                       [CRLF body]

The body is optional, and there is syntactic and semantic difference
between an empty body and a missing body.  Enhance purebred-email to
support parsing message types that do not have a body, or have an
optional body.

Also add a default implementation for EqMessage:

  default eqMessage :: (Eq a) => Message s a -> Message s a -> Bool

Fixes: https://github.com/purebred-mua/purebred-email/issues/35