header folding: filter out empty lines
The header wrapping will wrap a long word to the next line, even if
it is the first word. This is legal, but a bit awkward and
unnecessary.  It also complicates header processing e.g. when the
initial datum is an encoded-word value, the interpretation includes
spurious leading whitespace.

Filter out empty folded lines to avoid this problem.