RFC5322: add better date-time parsing
Add better parser, following strictly the specification from RFC
5322 section 3.3 "Date and Time Specification"[1]

[1]: https://tools.ietf.org/html/rfc5322#section-3.3

This change requires time >= 1.9, so bump the lower bound in the
package description.

Also, GHC < 8.8 bundles time 1.8.  This is not a fundamental
compatibility problem, but it requires annoying hacks for CI.  So
drop GHC 8.4 and 8.6 from the CI matrix.

Part of: https://github.com/purebred-mua/purebred-email/issues/49