ci: avoid 'installed' constraints
We now require 'time' >= 1.9.  GHC bundles 'time' itself, and
several packages that depend on 'time'.  GHC < 8.8 has 'time' < 1.9,
so the 'installed' constraint on these packages rules out a build
plan.  Remove the 'installed' constraint.

To keep it simple, remove the installed constraint from *all*
packages.  I don't think it's really a problem to depend on a newer
version of a bundled package, as long as there's a build plan and
GHC can actually build it!