mv Data.RFC5322 -> Data.IMF
Naming a module after a particular RFC is not wise.  RFC 2822 is,
after all, the third iteration of the Internet Message Format
standard (after RFC 2822 and RFC 822 before it).  And the IETF
emailcore working group is preparing the next version at this very
time!

Therefore move the modules from "Data.RFC5322" to "Data.IMF" (for
"Internet Message Format").  Retain the Data.RFC5322[.*] modules as
re-exports.  This commit does not deprecate data.RFC5322[.*], a
subsequent commit will do that.

Also do drive-by addition of the copyright and license comment on
the files touched by this commit.