always set Content-Type header for Multipart body
To ensure consistency, when rendering Multipart messages always
ensure the Content-Type header is set to a "multipart/" type with
the correct boundary parameter.

If the existing headers has a multipart Content-Type, just (re)set
the "boundary" parameter.  Otherwise, (re)set the Content-Type
header to multipart/mixed with the correct boundary.