improve text/* charset decoding framework
Some text/* subtypes specify that parameter value overrides charset
info carried in payload.  And for other subtypes, it is the other
way around.  In some cases, there is even a precedence of some
property of the payload, then the Content-Type parameter, and THEN
some other property of the payload.  Update the EntityCharsetSource
data type to support such nuanced charset selection.

Also improve the documentation of the `HasCharset ByteEntity`
instance.

Proper charset selection from text/xml and text/html remains a TODO.
of