avoid SMP-related MIN_VERSION_base CPP if-then-else
To gracefully handle the GHC Semigroup Monoid Proposal, we
introduced a constraint synonym with different implementations for
base < 4.11 and base >= 4.11.

We are already depending on base >= 4.11, so we can remove the CPP
if-then-else.  This also dispels an HLint parse failure (didn't like
the '#').

We could also ditch the `SM` constraint synonym but this is part of
the public API, so defer that change until a major version bump is
coming.