restrict multipart/related "type" param to type/subtype only
The "type" parameter of multipart/related Content-Type cannot have
any parameters.  Implement this restriction in the data type.

To minimise breakage and facilitate code reuse, change

    data ContentType = ContentType (CI B.ByteString) (CI B.ByteString) Parameters

into

    data ContentTypeWith a = ContentType (CI B.ByteString) (CI B.ByteString) a
    type ContentType = ContentTypeWith Parameters

The "bare content-type" case is represented as `ContentTypeWith ()`.

Ambiguity can arises in some cases.  For example:

    ContentType "text" "plain" mempty

Is `mempty` `()` or `Parameters []`?  To assist users in this case,
also provide:

    emptyParameters :: Parameters
    emptyParameters = mempty