support MIME subparts without bodies
Add support for parsing and constructing MIME parts without bodies.
The new `entities'` traversal visits all parts, projecting `Just
body` where there is a body, and `Nothing` where there is no body.
The `MIME` data type gets the new `PartNoBody` constructor.

This was a rather intrusive change.  Instead of passing around the
parser for "take a part at the current level", we pass around a new
data type that represents the current level:

    data MIMELevel = Top | Nest B.ByteString {- delimiter -}

We construct the "take part" parser on-demand where needed, and when
we are parsing terminal parts in a multipart message we avoid
preceding the delimiter with CRLF.  This lets us detect and handle
the "no body" case.

Fixes: https://github.com/purebred-mua/purebred-email/issues/90